src SwingDemo.java
This is a simple standard display. First its important to understand the JFrame object that creates the Window. We looked into the size and width  (new Dimension(200, 200)); and ect. THEN the main method must be static and that is apart of the Swing programs. Notice theres alot of hidden stuff within the window because I set it to visible or hidden as when needed to maintain some content along the way.